init: support dsds variant in one zip
The ONLY difference in blobs are the modem blobs. On devices that
have a dedicated modem partition supporting all variants is easy.
On other devices we can ship both sets of blobs. This adds less
than 30mb to our ROM and our system partition is huge.

We detect a DSDS device with some nice cmdline trickery. We use
the fact that "oemandroidboot.phoneid" is twice as long, then set
the correct props.

Change-Id: I3c043aaad841276032ee00b27c0de9ea0b2ae9e9
Signed-off-by: Adam Farden <adam@farden.cz>